backend/requirements.txt
Django==3.2.20
django-filter==23.2
djangorestframework==3.14.0
Pillow==10.0.0
isort==5.12.0
flake8==6.0.0
djoser==2.2.0
psycopg2-binary==2.9.6
django-colorfield==0.9.0
gunicorn==20.1.0
django-environ==0.10.0
